requirements.txt
requests
torchvision
scipy
scikit-image
aiohttp
pilgram
tkinter
base64
lzma